Separate TileLayer in to multiple files for readability
