+++ lib/src/layer/tile_builder/tile_builder.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_map/src/layer/tile_layer.dart';
+import 'package:flutter_map/src/layer/tile_layer/tile.dart';
 
 typedef TileBuilder = Widget Function(
     BuildContext context, Widget tileWidget, Tile tile);
