+++ lib/src/layer/tile_layer/coords.dart
@@ -1,0 +1,23 @@
+import 'dart:ui';
+
+import 'package:flutter_map/flutter_map.dart';
+
+class Coords<T extends num> extends CustomPoint<T> {
+  late T z;
+
+  Coords(T x, T y) : super(x, y);
+
+  @override
+  String toString() => 'Coords($x, $y, $z)';
+
+  @override
+  bool operator ==(Object other) {
+    if (other is Coords) {
+      return x == other.x && y == other.y && z == other.z;
+    }
+    return false;
+  }
+
+  @override
+  int get hashCode => hashValues(x.hashCode, y.hashCode, z.hashCode);
+}
