+++ lib/src/layer/tile_layer/level.dart
@@ -1,0 +1,9 @@
+import 'package:flutter_map/flutter_map.dart';
+
+class Level {
+  late double zIndex;
+  CustomPoint? origin;
+  late double zoom;
+  late CustomPoint translatePoint;
+  late double scale;
+}
