+++ lib/src/layer/tile_provider/file_tile_provider_io.dart
@@ -1,9 +1,9 @@
 import 'dart:io';
 
 import 'package:flutter/widgets.dart';
+import 'package:flutter_map/flutter_map.dart';
+import 'package:flutter_map/src/layer/tile_layer/coords.dart';
 
-import 'package:flutter_map/src/layer/tile_layer.dart';
-import 'package:flutter_map/src/layer/tile_provider/tile_provider.dart';
 
 /// FileTileProvider
 class FileTileProvider extends TileProvider {
