+++ lib/src/layer/tile_provider/file_tile_provider_web.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/widgets.dart';
-
-import 'package:flutter_map/src/layer/tile_layer.dart';
-import 'package:flutter_map/src/layer/tile_provider/tile_provider.dart';
+import 'package:flutter_map/flutter_map.dart';
+import 'package:flutter_map/src/layer/tile_layer/coords.dart';
 
 /// FileTileProvider
