+++ lib/src/layer/tile_provider/tile_provider.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_map/flutter_map.dart';
+import 'package:flutter_map/src/layer/tile_layer/coords.dart';
 
 import 'package:flutter_map/src/layer/tile_provider/network_image_with_retry.dart';
